requirements.txt
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
enum34==1.1.6
funcsigs==1.0.2
futures==3.1.1
html5lib==0.9999999
ipython==5.5.0
ipython-genutils==0.2.0
Keras==2.1.1
Markdown==2.6.9
matplotlib==2.1.0
mock==2.0.0
numpy==1.13.3
olefile==0.44
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.3.0
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
protobuf==3.5.0.post1
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
scandir==1.6
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
subprocess32==3.2.7
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0rc3
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.12.2
